requirements.txt
deepface>=0.0.79
numpy==1.26.3
scikit-learn>=0.24.0
tqdm>=4.64.0
pillow>=9.0.0
opencv-python>=4.5.0